+++ Lab-11-petclinic-solution/src/main/java/com/petclinic/core/VisitRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface VisitRepository extends JpaRepository<Visit, Integer> {
+public interface VisitRepository extends JpaRepository<Visit, Long> {
     public Optional<Visit> findByReferenceNumber(String referenceNumber);
 }
